move Task to Model
